Cropify_sih/requirement.txt
absl-py==0.9.0
asgiref==3.2.6
astor==0.8.1
backcall==0.2.0
bcrypt==3.1.7
cachetools==4.1.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
colorama==0.4.3
cryptography==3.0
decorator==4.4.2
Django==3.0.4
-e django-bcrypt==0.9.2
django-cors-headers==3.4.0
django-crispy-forms==1.9.0
django-rest-knox==4.1.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
dnspython==1.16.0
future==0.18.2
gast==0.2.2
geocoder==1.38.1
get==2019.4.13
google-auth==1.14.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.28.1
h5py==2.10.0
idna==2.9
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
joblib==0.16.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
locationiq==0.0.2
Markdown==3.2.1
mysqlclient==2.0.1
numpy==1.18.3
oauthlib==3.1.0
opt-einsum==3.2.1
pandas==1.0.3
parso==0.7.1
pickleshare==0.7.5
post==2019.4.13
prompt-toolkit==3.0.5
protobuf==3.11.3
public==2019.4.13
py-bcrypt==0.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.6.1
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2019.3
query-string==2019.4.13
ratelim==0.1.6
request==2019.4.13
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.23.1
scipy==1.4.1
six==1.14.0
sklearn==0.0
sqlparse==0.3.1
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
threadpoolctl==2.1.0
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.2.5
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
xlrd==1.2.0
